utility functions, remove _ms suffix
