+++ p2p/src/authenticated/actors/tracker/actor.rs
@@ -6,8 +6,8 @@ pub use super::{
 use crate::authenticated::{ip, metrics, wire};
 use bitvec::prelude::*;
 use commonware_cryptography::{PublicKey, Scheme};
-use commonware_runtime::{epoch, Clock, Spawner};
-use commonware_utils::{hex, union, DurationExt as _};
+use commonware_runtime::{Clock, Spawner};
+use commonware_utils::{hex, union, DurationExt as _, SystemTimeExt as _};
 use futures::{channel::mpsc, StreamExt};
 use governor::{
     clock::Clock as GClock, middleware::NoOpMiddleware, state::keyed::HashMapStateStore,
@@ -161,7 +161,7 @@ pub struct Actor<E: Spawner + Rng + GClock, C: Scheme> {
 impl<E: Spawner + Rng + Clock + GClock, C: Scheme> Actor<E, C> {
     pub fn new(runtime: E, mut cfg: Config<C>) -> (Self, Mailbox<E>, Oracle<E>) {
         // Construct IP signature
-        let timestamp = epoch(&runtime).as_millis_u64();
+        let timestamp = runtime.current().epoch_millis();
         let (socket_bytes, payload_bytes) = socket_peer_payload(&cfg.address, timestamp);
         let ip_namespace = union(&cfg.namespace, NAMESPACE_SUFFIX_IP);
         let ip_signature = cfg.crypto.sign(&ip_namespace, &payload_bytes);
@@ -428,7 +428,7 @@ impl<E: Spawner + Rng + Clock + GClock, C: Scheme> Actor<E, C> {
             }
 
             // If any timestamp is too far into the future, disconnect from the peer
-            let current_time = epoch(&self.runtime).as_millis_u64();
+            let current_time = self.runtime.current().epoch_millis();
             if peer.timestamp > current_time + self.synchrony_bound.as_millis_u64() {
                 return Err(Error::InvalidSignature);
             }
