+++ stream/src/placeholder/instance.rs
@@ -17,8 +17,8 @@ use chacha20poly1305::{
 };
 use commonware_cryptography::{PublicKey, Scheme};
 use commonware_macros::select;
-use commonware_runtime::{epoch, Clock, Sink, Spawner, Stream};
-use commonware_utils::DurationExt as _;
+use commonware_runtime::{Clock, Sink, Spawner, Stream};
+use commonware_utils::SystemTimeExt as _;
 use rand::{CryptoRng, Rng};
 
 // When encrypting data, an encryption tag is appended to the ciphertext.
@@ -49,7 +49,7 @@ impl<C: Scheme, Si: Sink, St: Stream> Instance<C, Si, St> {
         let ephemeral = x25519_dalek::PublicKey::from(&secret);
 
         // Send handshake
-        let timestamp = epoch(&runtime).as_millis_u64();
+        let timestamp = runtime.current().epoch_millis();
         let msg = create_handshake(
             &mut config.crypto,
             &config.namespace,
@@ -118,7 +118,7 @@ impl<C: Scheme, Si: Sink, St: Stream> Instance<C, Si, St> {
         let ephemeral = x25519_dalek::PublicKey::from(&secret);
 
         // Send handshake
-        let timestamp = epoch(&runtime).as_millis_u64();
+        let timestamp = runtime.current().epoch_millis();
         let msg = create_handshake(
             &mut config.crypto,
             &config.namespace,
